+++ tests/Stubs/Validator/Model.php
@@ -3,19 +3,17 @@ declare(strict_types=1);
 
 namespace Dogado\JsonApi\Server\Tests\Stubs\Validator;
 
-use Dogado\JsonApi\Annotations\Attribute;
-use Dogado\JsonApi\Annotations\Type;
+use Dogado\JsonApi\Attribute\Attribute;
+use Dogado\JsonApi\Attribute\Type;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @Type("validation-test")
- */
+#[Type('validation-test')]
 class Model
 {
-    /**
-     * @Attribute("jsonApiValueObject")
-     * @Assert\Valid()
-     */
+    #[
+        Attribute('jsonApiValueObject'),
+        Assert\Valid
+    ]
     private ValueObjectModel $valueObject;
 
     public function __construct()