+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.0.0 - 2021-07-XX
+
+- drop php 7.4 support
+- drop primary `doctrine/annotations` support for model validation
+- require `dogado/json-api-common:^2.0`
+- prepare support for `symfony/validator:^6.0`
+
+See [UPGRADE-2.0.md](UPGRADE-2.0.md) for more details.
+
 ## v1.0.0 - 2021-05-21
 
 - initial release